backend/requirements.txt
# FinIQ.ai - Python Dependencies

fastapi==0.115.0
uvicorn==0.30.6
# Align with broader ecosystem and satisfy libs needing >=2.8
pydantic==2.10.4
python-dotenv==1.0.0
redis==5.0.7
# Compatible with langchain-google-genai if present; works with our agents
google-generativeai==0.8.5
requests==2.32.3
